backend/requirements.txt
# ── Web Framework ──
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.22
slowapi==0.1.9

# ── Database ──
SQLAlchemy==2.0.46
psycopg==3.3.2
psycopg-binary==3.3.2
pgvector==0.3.6

# ── Supabase (Storage & Auth) ──
supabase==2.28.0
httpx==0.28.1

# ── Authentication ──
bcrypt==5.0.0
python-jose[cryptography]==3.5.0

# ── AI / NLP ──
sentence-transformers==5.2.2
torch==2.10.0
transformers==5.0.0
spacy==3.8.11
numpy==2.4.1
scikit-learn==1.8.0

# ── Document Parsing ──
PyPDF2==3.0.1
python-docx==1.1.2

# ── Configuration ──
pydantic-settings==2.13.0
python-dotenv==1.2.1
